requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.4
reportlab==4.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
